% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_input_df}
\alias{check_input_df}
\title{check_input_df()}
\usage{
check_input_df(df, expected_columns = NULL)
}
\arguments{
\item{df}{dataframe: dataframe}

\item{expected_columns}{character: If relevant, vector containing the names of the expected columns}
}
\value{
logical
}
\description{
Internal: helper function to wrap series of checks for dataframe inputs
}

