% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classify_question}
\alias{classify_question}
\title{#' classify_question.R()}
\usage{
classify_question(df, column = "weights")
}
\arguments{
\item{df}{dataframe: dataframe}

\item{column:}{the name of the column to be used for classification (column containing the weights)}
}
\value{
dataframe
}
\description{
Identifies whether questions are binary or likert types.
}

