% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_clean_survey.R
\name{clean_survey}
\alias{clean_survey}
\title{clean_survey()}
\usage{
clean_survey(survey, codebook, scale = NULL)
}
\arguments{
\item{survey}{dataframe: Dataframe containing survey data (survey tab)}

\item{codebook}{dataframe: Dataframe containing codebook data (codebook tab)}

\item{scale}{list: list of named vectors identifying questions to be aggregated into scales}
}
\value{
list
}
\description{
This function takes two dataframes containing raw ETO data as input, and formats them for further analysis. Both dataframes come from the  "[Admin] raw_prepost" report.
}
\examples{
df <- laycUtils::load_csv("./path/to/survey/data.csv")
codebook <- laycUtils::load_csv("./path/to/codebook/data.csv")
clean_survey(survey = df, codebook = codebook)
}

