% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\docType{data}
\name{codebook}
\alias{codebook}
\title{Codebook associated with fake survey data}
\format{Data frame with columns
\describe{
\item{pseudo}{Survey question short name}
\item{question}{Survey question}
\item{answer}{Possible answers to survey question}
\item{weight}{Weight associated to the answer}
}}
\source{
LAYC
}
\usage{
codebook
}
\description{
Codebook associated with fake survey data
}
\keyword{datasets}

