% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\docType{data}
\name{survey}
\alias{survey}
\title{Fake survey data}
\format{Data frame with columns
\describe{
\item{id}{Participants' unique id number}
\item{survey}{Name of the survey}
\item{response_id}{Survey response unique id number}
\item{question}{Survey question}
\item{pseudo}{Survey question short name}
\item{date}{Date the survey was taken}
\item{answer}{Answer to survey question}
\item{weight}{Weight associated to the answer}

}}
\source{
LAYC
}
\usage{
survey
}
\description{
Fake survey data
}
\keyword{datasets}

